json: show more descriptive error when JSON output is not available
The practice of calling exit_bad_args() on commandline parser errors is
convenient for developers but not for users as they get no hint which
argument is wrong and why. In particular, when JSON output is not available
for any reason, it may be hard to realize this is the cause rather than an
actual syntax error.

When JSON output is not available because it is not implemented or because
ioctl code path is used, tell user this is the problem.

Signed-off-by: Michal Kubecek <mkubecek@suse.cz>